Spike Aggregatwurzel: Refactor

+++ domaene/src/main/java/de/therapeutenkiller/dominium/lagerung/EreignisLager.java
@@ -20,9 +20,9 @@ public interface EreignisLager<E, A> {
     void ereignisseDemStromHinzufügen(
             String streamName,
             Collection<? extends Domänenereignis<A>> domänenereignisse,
-            Optional<Integer> erwarteteVersion);
+            Optional<Long> erwarteteVersion);
 
-    List<Domänenereignis<A>> getEreignisListe(String streamName, int vonVersion, int bisVersion);
+    List<Domänenereignis<A>> getEreignisListe(String streamName, long vonVersion, long bisVersion);
 
     void schnappschussHinzufügen(String streamName, E snapshot);
